Refactor project structure: reorganize scrapers and captcha solver modules
- Move selenium_scraper and requests_scraper to app/scrapers/ (each contains parser and fetch modules)
- Relocate captcha solver components (model, preprocessing, prediction) to app/captcha_solver/
- Add __init__.py files to new package directories for proper module imports
- Centralize course data: available and unavailable course JSONs now stored in app/courses_data/ (auto-creates if not exists)
- Move .env file to app/scrapers/ for shared configuration access between both scrapers
- Replace relative file paths with absolute paths to ensure data files are saved to correct location regardless of execution context
- Optimize captcha processing: pass image content directly to predict() instead of saving to disk and reading back
- Add enrollment_conditions field merging Enrollment Restriction and Entry Semester Eligibility
