--- app/.env
@@ -1,2 +1,0 @@
-USERNAME=your_student_id
-PASSWORD='your_password'
