+++ app/scrapers/.env
@@ -1,0 +1,2 @@
+USERNAME=your_student_id
+PASSWORD='your_password'
